chore: Remove AsyncClient and refactor extensions
